Share button to recipe page, User information endpoint
Added a share button to the recipe page, which  uses a function that copies the current URL to the clipboard. Added a endpoint to get all the necessary user information to display the profile page, still needs to complete the page with all information. Modified the query users only return needed information. Modified the gemini's prompt to send 3 to 5 recipes.